move some methods to rtc_helper.py
